+++ JS/Historique.log
@@ -1,50 +1,4 @@
 Le premier joueur commence
-Le joueur 1 a posé le mot CIEL
+Le joueur 1 a posé le mot DNS
 Le joueur 1 passe son tour
-Le joueur 2 ne Jarnac pas
-Le joueur 2 a pioché
-Le joueur 2 a posé le mot LESIU
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 passe son tour
-Le joueur 2 a Jarnac le mot CIELLL
-Le joueur 2 a pioché
-Le joueur 2 a posé le mot GUT
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 a posé le mot YAX
-Le joueur 1 passe son tour
-Le joueur 2 a Jarnac le mot YAXT
-Le joueur 2 a pioché
-Le joueur 2 a posé le mot ENS
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 passe son tour
-Le joueur 2 ne Jarnac pas
-Le joueur 2 a pioché
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 passe son tour
-Le joueur 2 ne Jarnac pas
-Le joueur 2 a pioché
-Le joueur 2 a posé le mot NYU
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 a posé le mot AOT
-Le joueur 1 passe son tour
-Le joueur 2 a Jarnac le mot NAOT
-Le joueur 2 a pioché
-Le joueur 2 passe son tour
-Le joueur 1 ne Jarnac pas
-Le joueur 1 a pioché
-Le joueur 1 passe son tour
-Le joueur 2 ne Jarnac pas
-Le joueur 2 a pioché
-Le joueur 2 a posé le mot VIN
-La partie est terminée
-Le joueur 3 a gagné
+Le joueur 2 a Jarnac le mot DENS
